refactor for modular and reusable imp
